chore: empty README added to backend
