+++ moutarde-back/moutarde-back.sln
@@ -1,6 +1,6 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "moutarde-back", "moutarde-back.csproj", "{A744DF4B-CBAD-436E-A6E7-7B6D85A1CF48}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "moutarde-back", "moutarde-back\moutarde-back.csproj", "{64AEE6DE-2D87-417C-B9C6-6EBD96C19FA3}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -8,9 +8,9 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A744DF4B-CBAD-436E-A6E7-7B6D85A1CF48}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A744DF4B-CBAD-436E-A6E7-7B6D85A1CF48}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A744DF4B-CBAD-436E-A6E7-7B6D85A1CF48}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A744DF4B-CBAD-436E-A6E7-7B6D85A1CF48}.Release|Any CPU.Build.0 = Release|Any CPU
+		{64AEE6DE-2D87-417C-B9C6-6EBD96C19FA3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{64AEE6DE-2D87-417C-B9C6-6EBD96C19FA3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{64AEE6DE-2D87-417C-B9C6-6EBD96C19FA3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{64AEE6DE-2D87-417C-B9C6-6EBD96C19FA3}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
